pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finance-data-platform"
version = "0.1.0"
description = "Layered data platform for financial market data."
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Shervin Taheripour" },
]
dependencies = [
  "pydantic>=2.6,<3.0",
  "PyYAML>=6.0,<7.0",
  "pandas>=2.2,<3.0",
  "yfinance>=0.2.50,<1.0",
  "pyarrow>=16.0,<21.0",
  "duckdb>=1.0,<2.0",
  "Jinja2>=3.1,<4.0",
  "matplotlib>=3.8,<4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "ruff>=0.11,<1.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "N", "W"]
